+++ mvnw.cmd
@@ -1,4 +1,4 @@
-<# : batch portion
+build.gradle<# : batch portion
 @REM ----------------------------------------------------------------------------
 @REM Licensed to the Apache Software Foundation (ASF) under one
 @REM or more contributor license agreements.  See the NOTICE file
