+++ pom.xml
@@ -58,6 +58,7 @@
 			<artifactId>google-api-client</artifactId>
 			<version>1.35.2</version>
 		</dependency>
+
 		<dependency>
 			<groupId>net.dv8tion</groupId>
 			<artifactId>JDA</artifactId>
@@ -69,12 +70,14 @@
 			<artifactId>spring-boot-configuration-processor</artifactId>
 			<optional>true</optional>
 		</dependency>
+
 		<dependency>
 			<groupId>org.projectlombok</groupId>
 			<artifactId>lombok</artifactId>
 			<optional>true</optional>
 		</dependency>
-		<dependency>
+
+			<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-test</artifactId>
 			<scope>test</scope>
@@ -92,7 +95,7 @@
 		</dependency>
 
 		<dependency>
-			<groupId>com.h2database</groupId> <!-- ou MySQL/PostgreSQL/etc -->
+			<groupId>com.h2database</groupId>
 			<artifactId>h2</artifactId>
 			<scope>runtime</scope>
 		</dependency>
