+++ src/main/java/brawl/example/brawl_tracker/brawl_core/entity/SaveMongo.java
@@ -1,4 +1,4 @@
-package brawl.example.brawl_tracker.entity;
+package brawl.example.brawl_tracker.brawl_core.entity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
