+++ src/main/java/brawl/example/brawl_tracker/brawl_core/model/PlayerValidModel.java
@@ -1,4 +1,4 @@
-package brawl.example.brawl_tracker.model;
+package brawl.example.brawl_tracker.brawl_core.model;
 
 import lombok.*;
