+++ src/main/java/brawl/example/brawl_tracker/brawl_core/service/BrawlService.java
@@ -1,8 +1,8 @@
-package brawl.example.brawl_tracker.service;
+package brawl.example.brawl_tracker.brawl_core.service;
 
-import brawl.example.brawl_tracker.dto.TeamBattleDTO;
-import brawl.example.brawl_tracker.model.BrawlRequestMODEL;
-import brawl.example.brawl_tracker.model.TeamMODEL;
+import brawl.example.brawl_tracker.brawl_core.dto.TeamBattleDTO;
+import brawl.example.brawl_tracker.brawl_core.model.BrawlRequestMODEL;
+import brawl.example.brawl_tracker.brawl_core.model.TeamMODEL;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.*;
