+++ src/main/java/brawl/example/brawl_tracker/brawl_core/util/OpenURL.java
@@ -1,4 +1,4 @@
-package brawl.example.brawl_tracker.util;
+package brawl.example.brawl_tracker.brawl_core.util;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.web.client.RestTemplate;
